pyproject.toml
[project]
name = "ai"
version = "0.1.0"
description = "RAG on a stick"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.1.0",
    "langchain>=0.3.27",
    "langchain-chroma>=0.2.6",
    "langchain-community>=0.3.30",
    "langchain-huggingface>=0.3.1",
    "langchain-ollama>=0.3.10",
    "notebook>=7.4.7",
    "sentence-transformers>=5.1.1",
    "unstructured[pdf]>=0.18.15",
]
